Starting Code for Day 7 Incomplete
Started work with parsing stuff still debating on how to approach this. 
Might need to do some sort of recursive solution to calculate the value of the input lines. In any case it is a matter of dilineating between binary or unary operations, having that instruction with src and destination sorted out. Have some sort of conditional to parse out the desired code implementation and then doing something like that. The main challenge is that things aren't really concretely defined, you have a bunch of operations going on in the direction list but the values aren't stated. So then you have to actually calculate values from the limited that are given explicity which seems to be two for my input prompt and then work through with that limited operations to fill out from.